feat: fetch settings
